Add auto-importer for custom tags
To use it, you have to add the import to components/index.tsx so it is exposed.

We could use dynamic imports from webpack or glob over all the files in that directory when the server starts up but that does not seem worth the effort.

Personally, I do not think this PR is adding much value but putting it up for consideration