Improve image support
This adds support for an img tag that passes through to the next.js Image
component. It also updates the figure tag to use the Image component and provides
some default styling that I think matches common use cases. If nothing else, it
should provide the scaffolding to make it work throughout the docs.

It is possible to override the default img renderer but I did not do that because
next.js images require height/width or fill to handle responsive images and the
default behavior in markdoc seems like a better default for now.