Fix link underlines #168
I noticed during my audit that the underlines are different depending on
if you are in light mode or dark. Here's the before:

https://github.com/replayio/docs/assets/5903784/b361c4f7-543a-4060-ba9f-1cf7f4215af5

Here's my attempt at improving it:

https://github.com/replayio/docs/assets/5903784/fb5e83d7-0818-4954-b93a-97f26a7f7ee8

This is now consistent at least. I have also bumped the hover state down
in size. It's still a significant enough difference that you can
definitely tell the hovered one apart, but it doesn't overlap the text
as much. I found hovered links difficult to read in the previous
iteration, but I'm open to feedback about that piece of the change.